backend/requirements.txt
# FastAPI + ASGI server
fastapi==0.128.0
uvicorn==0.40.0

# LangChain for RAG pipeline
langchain==1.2.7
langchain.community==0.4.1
langchain-text-splitters>=1.1.0
langchain-openai

# OpenAI SDK
openai

# FAISS vectorstore (CPU version)
faiss-cpu==1.13.2

# Embeddings / NLP
tiktoken==0.12.0

# PDF / text loaders
PyMuPDF==1.26.7
pdfminer.six==20260107

# Document splitting
python-docx==1.2.0

# Data handling
numpy==2.4.2
pandas==3.0.0

# Environment variables
python-dotenv==1.2.1

# Pydantic for typed schemas
pydantic==2.12.5

# Regex / security scanning
regex==2026.1.15

# Optional: Tree-sitter (if later added AST analysis)
tree_sitter==0.25.2 # Or latest version when added

# HTTP requests (if needed for extra docs)
requests==2.32.5
